simple-pipecat-webrtc/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
aiortc==1.5.0
pipecat-ai[webrtc]